Break scheduling fully functional

+++ app/src/main/java/com/smutkiewicz/blinkbreak/model/Task.kt
@@ -1,4 +1,4 @@
 package com.smutkiewicz.blinkbreak.model
 
-data class Task(val breakEvery: Int, val breakDuration: Int,
+data class Task(val breakEvery: Long, val breakDuration: Long,
                 val areNotificationsEnabled: Boolean, val isLowerBrightness: Boolean)
